requirements.txt
undetected-chromedriver==3.5.5
selenium==4.15.2
pyTelegramBotAPI==4.22.1
requests==2.32.3
python-dotenv==1.0.1